So much was added.
It finally loads Feeds and everything seem to be clickable for now.
Yet to be added is image loading functionality, still unsure how that
works, and also the ability to go directly to the webpage of both the
entry and the feed source page.

I promise to make more commits in the future.